fized overflowing issues
