+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:cropapp/Pages/adminConsole.dart';
+import 'package:cropapp/Screens/adminConsole.dart';
 import 'package:flutter/material.dart';
 
 void main() {
